+++ src/main/java/nomina/soft/backend/services/AfpService.java
@@ -15,7 +15,7 @@ public interface AfpService {
 	
 	public AfpModel guardarAFP(AfpDto afpDto) throws AfpNotFoundException, AfpExistsException;
 	
-	public AfpModel updateAfp(String actualNombre, String nombre, Double actualDescuento, Double descuento) throws AfpNotFoundException, AfpExistsException;
+	public AfpModel updateAfp(String actualNombre, String nombre, float actualDescuento, float descuento) throws AfpNotFoundException, AfpExistsException;
 	
 	public void deleteAfp(String nombre);
